+++ test/suite/controller/fileDecoratorController.test.ts
