+++ test/suite/controller/statusBarController.test.ts
