+++ test/suite/model/fileDecorationProvider.test.ts
